Extract ItemsLookup class to separate file.
