+++ test/lib/coloration/support/items_lookup_test.rb
@@ -1,0 +1,15 @@
+require 'test_helper'
+
+module Coloration
+
+  module Readers
+
+    describe ItemLookup do
+
+      let(:described) { Coloration::Readers::ItemLookup }
+
+    end # ItemLookup
+
+  end # Readers
+
+end # Coloration
